+++ CHANGES.rst
@@ -15,6 +15,8 @@ Improvements
 - Show document templates from non-parent categories and other events for cloning
   as long as the user has management access (:pr:`6232`)
 - Warn about conflicts from concurrent edits of minutes (:issue:`3410`, :pr:`6193`)
+- Include up to two months (up from one week) of past events in dashboard iCal export
+  (:pr:`6304`)
 
 Bugfixes
 ^^^^^^^^